feat(core): 修改分支信息中的修改时间数据类型
将BranchInfo类中的modified_time字段从int类型改为float类型，
以支持更精确的时间戳存储。

+++ core/Data_Manager/User_Data_Manager/Sub_Manager/_branch_info.py
@@ -19,7 +19,7 @@ class BranchInfo(BaseModel):
     """Branch Info"""
     branch_id: str = ""
     size: int = 0
-    modified_time: int = 0
+    modified_time: float = 0
     
     @property
     def readable_size(self) -> str:
